[+] [MemoryExplorer]..
Well, I need to think about place for such code.

There is no specific hardware beside Uno. It's like demo of the library
framework. But hey, even demos are not libraries, they are standalone
programs!

And this one has one-job mission which is not demonstrate library
capabilities but dump memory.